[WFLY-266] Notification Support
moved notification-related result handlers to
NotificationResultHandlers
