+++ controller/src/main/java/org/jboss/as/controller/AbstractWriteAttributeHandler.java
@@ -24,12 +24,9 @@ package org.jboss.as.controller;
 
 import static org.jboss.as.controller.ControllerLogger.MGMT_OP_LOGGER;
 import static org.jboss.as.controller.ControllerMessages.MESSAGES;
-import static org.jboss.as.controller.OperationContext.ResultAction.KEEP;
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.ATTRIBUTE_VALUE_WRITTEN_NOTIFICATION;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.NAME;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.VALUE;
-import static org.jboss.as.controller.operations.global.GlobalNotifications.NEW_VALUE;
-import static org.jboss.as.controller.operations.global.GlobalNotifications.OLD_VALUE;
+import static org.jboss.as.controller.notification.NotificationResultHandlers.AttributeValueWrittenResultHandler;
 
 import java.util.Collection;
 import java.util.HashMap;
@@ -37,7 +34,6 @@ import java.util.Map;
 import java.util.concurrent.atomic.AtomicReference;
 
 import org.jboss.as.controller.descriptions.ModelDescriptionConstants;
-import org.jboss.as.controller.notification.Notification;
 import org.jboss.as.controller.operations.validation.ParametersValidator;
 import org.jboss.as.controller.registry.Resource;
 import org.jboss.dmr.ModelNode;
@@ -133,24 +129,7 @@ public abstract class AbstractWriteAttributeHandler<T> implements OperationStepH
             }, OperationContext.Stage.RUNTIME);
         }
 
-        context.completeStep(new OperationContext.ResultHandler() {
-            @Override
-            public void handleResult(OperationContext.ResultAction resultAction, OperationContext context, ModelNode operation) {
-                if (resultAction != KEEP) {
-                    return;
-                }
-                PathAddress sourceAddress = PathAddress.pathAddress(operation.get(ModelDescriptionConstants.OP_ADDR));
-                ModelNode data = new ModelNode();
-                data.get(NAME).set(attributeName);
-                data.get(OLD_VALUE).set(currentValue);
-                data.get(NEW_VALUE).set(newVal.get());
-                Notification notification = new Notification(ATTRIBUTE_VALUE_WRITTEN_NOTIFICATION,
-                        sourceAddress,
-                        MESSAGES.attributeValueWritten(attributeName, currentValue, newVal.get()),
-                        data);
-                context.emit(notification);
-            }
-        });
+        context.completeStep(new AttributeValueWrittenResultHandler(attributeName, currentValue, newVal.get()));
     }
